inle_17

+++ problem_0014/inle_14/inle_14.cpp
@@ -20,7 +20,7 @@ int length_collatz(int number)
 }
 
 
-int longest_collatz(int max) 
+int longest_collatz(int64_t max) 
 {
     int length;
     int longest;
